Implemented new interface classes to make my life easier
Made an "IInteractable" interface so I can classify other
objects as interactable other than items.

Made an "IUniqueEditorElement" interface so I could edit values
in a class that needed to be update other vaules in the class
based on what was changed.

Created an Item box spawner that will spawn item in either
sets of 4, 2, or 1 with 1 also being a legendary

Moved weapon stats generation to a new class and updated
the stats generation settings with more control on how to generate
the weapons